+++ sbk-api/src/main/java/io/sbk/api/impl/MetricsLogger.java
@@ -95,8 +95,8 @@ public class MetricsLogger implements Print {
 
 
     @Override
-    public void print(long bytes, long records, double recsPerSec, double mbPerSec, double avgLatency, int maxLatency,
-                      long invalid, long lowerDiscard, long higherDiscard, int[] percentileValues) {
+    public void print(long bytes, long records, double recsPerSec, double mbPerSec, double avgLatency, long maxLatency,
+                      long invalid, long lowerDiscard, long higherDiscard, long[] percentileValues) {
         this.bytes.increment(bytes);
         this.records.increment(records);
         this.invalidLatencyRecords.increment(invalid);
